+++ cpas_toolbox/config/cass.yaml
@@ -1,5 +1,5 @@
 name: CASS
-wrapper_type: cpas_toolbox.method_wrappers.CASSWrapper
+wrapper_type: cpas_toolbox.cpas_methods.CASS
 config_dict:
   num_points: 500  # points in reconstructed pointcloud
   num_objects: 6  # number of categories
