+++ cpas_toolbox/cpas_methods/_spd/spd/align.py
